requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
torch
torchvision
pillow
